% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESizeCoHedg.dnsty02.R
\name{ESizeCoHedg.dnsty02}
\alias{ESizeCoHedg.dnsty02}
\title{Calculating the Effect Size of Cohen's d & Hedges'g with the density curve and mean's line}
\usage{
ESizeCoHedg.dnsty02()
}
\value{
Effect Size of Cohen's d & Hedges'g
}
\description{
\code{ESizeCoHedg} calculate the Effect Size & draw the density curve and mean's line
}
\examples{
# ESizeCoHedg.dnsty02()
}
